requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --annotation-style=line --index-url=https://pypi.org/simple --no-emit-index-url --output-file=requirements.txt requirements-in.txt
#
adbutils==0.11.0          # via -r requirements-in.txt, uiautomator2
aiofiles==23.1.0          # via -r requirements-in.txt
alas-webapp==0.3.7.0      # via -r requirements-in.txt
annotated-types==0.7.0    # via pydantic
antlr4-python3-runtime==4.9.3  # via omegaconf
anyio==3.6.2              # via -r requirements-in.txt, starlette, watchfiles
apkutils2==1.0.0          # via adbutils
cached-property==1.5.2    # via uiautomator2
certifi==2022.12.7        # via requests
cffi==2.0.0               # via gevent
chardet==5.1.0            # via -r requirements-in.txt
charset-normalizer==3.0.1  # via requests
cigam==0.0.3              # via apkutils2
click==8.1.3              # via uvicorn
colorama==0.4.6           # via click, colorlog, logzero, rich, tqdm, uvicorn
colorlog==6.10.1          # via rapidocr
commonmark==0.9.1         # via rich
contourpy==1.3.3          # via matplotlib
cycler==0.12.1            # via matplotlib
decorator==5.1.1          # via retry
deprecated==1.2.13        # via uiautomator2
deprecation==2.1.0        # via adbutils
filelock==3.9.0           # via uiautomator2
flatbuffers==25.12.19     # via onnxruntime-directml
fonttools==4.62.1         # via matplotlib
future==0.18.3            # via zerorpc
gevent==25.9.1            # via zerorpc
greenlet==3.3.2           # via gevent
h11==0.14.0               # via uvicorn
httptools==0.7.1          # via uvicorn
idna==3.4                 # via anyio, requests
imageio==2.26.0           # via -r requirements-in.txt
inflection==0.5.1         # via -r requirements-in.txt
jellyfish==0.9.0          # via -r requirements-in.txt
kiwisolver==1.5.0         # via matplotlib
llvmlite==0.46.0          # via numba
logzero==1.7.0            # via uiautomator2
lxml==6.0.2               # via uiautomator2
lz4==4.4.5                # via -r requirements-in.txt
matplotlib==3.10.8        # via -r requirements-in.txt
mpmath==1.3.0             # via sympy
msgpack==1.0.4            # via zerorpc
numba==0.64.0             # via -r requirements-in.txt
numpy==2.4.3              # via -r requirements-in.txt, contourpy, imageio, matplotlib, numba, onnxruntime-directml, opencv-python, rapidocr, scipy, shapely
omegaconf==2.3.0          # via rapidocr
onepush==1.2.0            # via -r requirements-in.txt
onnxruntime-directml==1.24.4  # via -r requirements-in.txt
opencv-python==4.13.0.92  # via -r requirements-in.txt, rapidocr
packaging==20.9           # via deprecation, matplotlib, onnxruntime-directml, uiautomator2
pillow==12.1.1            # via -r requirements-in.txt, imageio, matplotlib, rapidocr, uiautomator2
progress==1.6             # via uiautomator2
protobuf==6.33.5          # via onnxruntime-directml
psutil==5.9.4             # via -r requirements-in.txt
py==1.11.0                # via retry
pyclipper==1.4.0          # via rapidocr
pycparser==2.21           # via cffi
pycryptodome==3.23.0      # via -r requirements-in.txt
pydantic==2.12.5          # via -r requirements-in.txt
pydantic-core==2.41.5     # via pydantic
pyelftools==0.29          # via apkutils2
pygments==2.14.0          # via rich
pyparsing==3.0.9          # via matplotlib, packaging
pypresence==4.2.1         # via -r requirements-in.txt
python-dateutil==2.9.0.post0  # via matplotlib
python-dotenv==1.0.0      # via uvicorn
pywebio==1.7.1            # via -r requirements-in.txt
pyyaml==6.0.3             # via -r requirements-in.txt, omegaconf, rapidocr, uvicorn
pyzmq==27.1.0             # via -r requirements-in.txt, zerorpc
rapidocr==3.7.0           # via -r requirements-in.txt
requests==2.28.2          # via adbutils, onepush, rapidocr, uiautomator2
retry==0.9.2              # via adbutils, uiautomator2
retrying==1.3.4           # via -r requirements-in.txt
rich==11.2.0              # via -r requirements-in.txt
scipy==1.17.1             # via -r requirements-in.txt
shapely==2.1.2            # via rapidocr
six==1.16.0               # via adbutils, python-dateutil, rapidocr, retrying, uiautomator2
sniffio==1.3.0            # via anyio
starlette==0.25.0         # via -r requirements-in.txt
sympy==1.14.0             # via onnxruntime-directml
tornado==6.2              # via pywebio
tqdm==4.67.3              # via rapidocr
typing-extensions==4.15.0  # via -r requirements-in.txt, pydantic, pydantic-core, typing-inspection
typing-inspection==0.4.2  # via pydantic
ua-parser==0.16.1         # via user-agents
uiautomator2==2.16.17     # via -r requirements-in.txt
uiautomator2cache==0.3.0.1  # via -r requirements-in.txt
urllib3==1.26.14          # via requests
user-agents==2.2.0        # via pywebio
uvicorn[standard]==0.20.0  # via -r requirements-in.txt
watchdog==6.0.0           # via -r requirements-in.txt
watchfiles==0.18.1        # via uvicorn
websockets==10.4          # via uvicorn
whichcraft==0.6.1         # via adbutils, uiautomator2
wrapt==1.15.0             # via -r requirements-in.txt, deprecated
xmltodict==0.13.0         # via apkutils2
zerorpc==0.6.3            # via -r requirements-in.txt
zope-event==4.6           # via gevent
zope-interface==5.5.2     # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
